+++ R/etNew.R
@@ -407,6 +407,7 @@
   rownames(.out) <- seq_len(nrow(.out))
   attr(.out, "rxEtPreviewGroups") <- .meta
   attr(.out, "rxEtShow") <- envRef$show
+  attr(.out, "rxEtExtraCols") <- .etExtraCols(envRef) # nolint
   class(.out) <- c("rxEtPreview", class(.out))
   .tu <- envRef$units["time"]
   if (!is.na(.tu) && nchar(.tu) > 0 && requireNamespace("units", quietly = TRUE)) {